Fixing navigation and styles with bootstrap navbar component
