+++ src/components/Navigation.svelte
@@ -1,0 +1,35 @@
+<script>
+  import {
+    Collapse,
+    Navbar,
+    NavbarToggler,
+    NavbarBrand,
+    Nav,
+    NavItem,
+    NavLink
+  } from 'sveltestrap/src';
+  import Helth from '../components/Helth.svelte';
+
+  let isOpen = false;
+  const toggle = () => (isOpen = !isOpen);
+</script>
+
+<Navbar color="dark" dark>
+    <NavbarBrand class="me-auto" href="/">
+        <Helth />
+    </NavbarBrand>
+    <NavbarToggler on:click={toggle} class="me-2" />
+    <Collapse {isOpen} navbar>
+        <Nav navbar class="sticky-top">
+            <NavItem>
+                <NavLink href="/">Today</NavLink>
+            </NavItem>
+            <NavItem>
+                <NavLink href="/">History</NavLink>
+            </NavItem>
+            <NavItem>
+                <NavLink href="/">Profile</NavLink>
+            </NavItem>
+        </Nav>
+    </Collapse>
+</Navbar>
